fix conditionDisease
